move ErrorBoundary to components folder
